refactor: Output 을 테스트할 수 있도록 MockDrawView 작성, Controller 는 OutputFormat 을 받는 것으로 변경
