+++ src/main/kotlin/study/racinggame/presentation/view/DrawView.kt
@@ -1,21 +1,21 @@
 package study.racinggame.presentation.view
 
-import study.racinggame.domain.CarResponse
+import study.racinggame.presentation.dto.CarResponse
 
-class DrawView {
+class DrawView : OutputFormat {
 
     init {
         println("실행 결과")
     }
 
-    fun drawCarState(carResponses: List<CarResponse>) {
+    override fun drawCarState(carResponses: List<CarResponse>) {
         for (response in carResponses) {
             println(response.getPositionRepresentation())
         }
         println()
     }
 
-    fun drawRaceWinner(winnerName: String) {
+    override fun drawRaceWinner(winnerName: String) {
         println("${winnerName}님이 최종 우승했습니다.")
     }
 }
