+++ src/main/kotlin/study/racinggame/presentation/view/OutputFormat.kt
@@ -1,0 +1,8 @@
+package study.racinggame.presentation.view
+
+import study.racinggame.presentation.dto.CarResponse
+
+interface OutputFormat {
+    fun drawCarState(carResponses: List<CarResponse>)
+    fun drawRaceWinner(winnerName: String)
+}
