+++ src/test/kotlin/study/racinggame/presentation/RaceControllerTest.kt
@@ -2,8 +2,8 @@ package study.racinggame.presentation
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
+import study.racinggame.service.mock.MockDrawView
 import study.racinggame.service.mock.MockInputReader
-import study.racinggame.service.mock.MockRandom
 import java.io.ByteArrayOutputStream
 import java.io.PrintStream
 
@@ -19,9 +19,8 @@ class RaceControllerTest : StringSpec({
     }
 
     "컨트롤러가 정상적으로 등록되고, 작동에 에러가 없다." {
-        val mockInputReader = MockInputReader(listOf("park,sejun", "1"))
-        val mockRandom = MockRandom(Int.MAX_VALUE)
-        val controller = RaceController(mockInputReader, mockRandom)
+        val mockInputReader = MockInputReader(listOf("park,sejun", "2"))
+        val controller = RaceController(mockInputReader, MockDrawView())
         controller.startRace()
 
         val normalizedOutput = outputStreamCaptor.toString().trim().replace("\r\n", "\n")
@@ -31,10 +30,9 @@ class RaceControllerTest : StringSpec({
         |경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).
         |시도할 횟수는 몇 회인가요?
         |
-        |park : --
-        |sejun : --
-        |
-        |park, sejun님이 최종 우승했습니다.
+        |drawCarState is Called
+        |drawCarState is Called
+        |drawRaceWinner is Called
         """.trimMargin()
     }
